Add existing file

+++ timer/views.py
@@ -1,34 +1,34 @@
-from django.http import Http404
 from django.shortcuts import get_object_or_404, render, reverse
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.urls import reverse
+from django.contrib.auth.decorators import login_required
 
 from timer.models import Activity
-from timer.forms import ActivityForm
+from timer.forms import ActivityForm, ActivityFormEdit
 
 def index(request):
     return render(request, 'timer/index.html') 
-   
+
+@login_required   
 def activities(request):
     
-    activities = Activity.objects.order_by('-activity_time')[:]
+    activities = Activity.objects.filter(owner=request.user).order_by('-activity_time')[:]
     context = {'activities':activities}
     #the render() function takes the rquest object as its first argument, a template name as its second
     #argument and a dictionary as its optional third argument. It returns an HttpResponse object of the given
     #template rendered with the given context
     return render(request, 'timer/activities.html', context)
 
+@login_required 
 def activity_detail(request, activity_id):
     
-    #the get_object_or_404 takes a django model as its first argument and an arbitrary number of keyword arguments, 
-    #which it passe sto the get() function of the model's manager. It raises Http404 if the object doesnt exist
-    #activity = get_object_or_404(Activity, pk=activity_id)
     activity = Activity.objects.get(id=activity_id)
-    
-    
-    
+    #make sure the topic belonds to the curent user
+    if activity.owner != request.user:
+        raise Http404
     return render(request, 'timer/activity_detail.html', {'activity': activity})
     
+@login_required 
 def new_activity(request):
     """Add a new activity"""
     if request.method != 'POST': 
@@ -38,7 +38,10 @@ def new_activity(request):
         #POST data submitted; process data
         form = ActivityForm(request.POST)  
         if form.is_valid():   
-            form.save()    
+            new_activity = form.save(commit=False)
+            new_activity.owner = request.user
+            new_activity.save()
+            
             
             #activity_title = form.cleaned_data['activity_title']
             #activity = Activity.objects.filter(activity_title=activity_title)
@@ -47,17 +50,19 @@ def new_activity(request):
     context = {'form': form}   
     return render(request, 'timer/new_activity.html', context)
 
-
+@login_required 
 def edit_activity(request, activity_id):
     """Edit an existing activity"""
     activity = Activity.objects.get(id=activity_id)
+    if activity.owner != request.user:
+        raise Http404
     
     if request.method != 'POST':
         #initial request; pre-fill form with current entry
-        form = ActivityForm(instance=activity)
+        form = ActivityFormEdit(instance=activity)
     else:
         # POST data submitted; process data
-        form = ActivityForm(instance=activity, data=request.POST)
+        form = ActivityFormEdit(instance=activity, data=request.POST)
         if form.is_valid:
             form.save()
             return HttpResponseRedirect(reverse('timer:activity_detail', args=[activity.id]))
@@ -65,9 +70,11 @@ def edit_activity(request, activity_id):
     context = {'activity': activity, 'form': form}
     return render(request, 'timer/edit_activity.html', context)
 
-
+@login_required 
 def delete_activity(request, activity_id):
     """delete an existing activity"""
     activity = Activity.objects.get(id=activity_id)
+    if activity.owner != request.user:
+        raise Http404    
     activity.delete()
     return HttpResponseRedirect(reverse('timer:activities'))
